Check and accept EULA in brkt-cli
Add the following options at the top level of the brkt command:

--username (required)
--password (optional)
--api-root (hidden, development only)
--no-verify-cert (hidden, development only)

brkt-cli now authenticates with the Bracket service before doing
any work.

If --password is not specified, prompt the user for the password.
If the EULA has not been signed, prompt the user with the link to the
EULA text.  The user must type the word "YES" (case-insensitive) in
order to accept the EULA.

Add a new BracketService class for code that talks to the Bracket
service API.