Improve error handling during cleanup
If something unexpected happens during cleanup, log the traceback, so
that we know where the problem occurred.  Move the bulk of the cleanup
code into its own function, to avoid a massive finally block inside the
encrypt() function.

When deleting a volume, ignore InvalidVolume.NotFound.  This can happen
when AWS deletes a volume that was attached to a terminated instance,
and we then attempt to delete the volume explicitly.

Remove extra log message that was printed when encryption completes.