Specify subnet and security group for image encryption
Add --subnet and --security-group options to the encrypt-ami subcommand.
The subnet is used when running both the snapshotter and encryptor
instances.  The security groups are used when running the encryptor
instance.

Clean up the command line parsing code to check the subparser name
instead of referencing argv[0].  The old code broke when passing -v to
the brkt top-level command.