Always terminate guest instance
Move the wait_for_instance() call out of run_guest_instance(). If
anything goes wrong while we're waiting for the guest instance to start,
we will clean up the instance in the finally block.

Remove the updater argument from run_guest_instance. I believe this is a
vestige of an earlier version of the updater code, and is no longer
used.