Check version on startup
When starting brkt-cli, check the code version against the list of
available versions on PyPI.  Exit if the code version is earlier than
the earliest available version.  If a newer version is available, log a
message that reminds the user to upgrade.  Add a --no-check-version
option to skip the version check.