Pass --brkt-env value when updating AMI
Parse the --brkt-env value and add it to userdata when updating an
encrypted AMI.  Break out the --brkt-env parsing logic into a separate.
Pass bracket environment data around as an object instead of as a
string.  This allows us to use the parsing code for both encryption and
updating, and also add unit tests.

Move the hostname validation function into util, so that it can be used
to validate the hostnames specified in the --brkt-env value.