Compress userdata for encryptor instance
The encryptor AMI only looks for multi-part MIME userdata in
the case that it finds the userdata to be compressed.  So we'll
compress it, which is probably a good thing to do anyway.
Also, make the test_brkt_env_encrypt() test expect the userdata
to be in compressed MIME format, rather than plain JSON.