Use NO_PROXY talking to private encryptor ip
In some customer envrionments, customers use http_proxy
environment variables in the shell they are encrypting images
and it is needed for the cli when validating version. This
change will ensure we don't try to communicate to encryptor
instances with private ips through the proxy by appending the
encryptor private ip to the NO_PROXY environment variable.