Support specifying the guest instance type
Add a --guest-instance-type option to the encrypt-ami and update-
encrypted-ami subcommands.  This allows the user to specify the guest
instance type when the default of m3.medium is incompatible with the
guest AMI.

Clean up the original test_instance_type() test, so that it switches on
the AMI ID instead of making assumptions about the order in which
instances are launched.