Add proxy support to update-encrypted-ami
Add --proxy and --proxy-config-file options to the update-encrypted-ami
subcommand.

Add a _get_proxy_config() utility function for loading proxy.yaml
content.  This function is called during encryption and update.

Add a combine_user_data() function, which generates the gzipped MIME
blob that is passed to the metavisor.  This allows us to write dedicated
unit tests for this logic.

Minor terminology changes:
* user data: the actual user data bytes that are sent to the instance
* brkt env: host/port of the api service and HSM proxy
* brkt config: the dictionary that stores brkt env, ntp servers, etc.