Fix disk 404 issues
So it turns out google is doing some serious
shenanigans on the backend. When I tracked the
api requests more carefully I saw disks and
instances completely disappearing for extended
periods of time.

The only place I (consistently) saw this happening
was the guest instance and root disk. The guest root
disk is 'created' by launching an instance from
the guest image and immediately deleting the instance
when it became ready and attaching that guest root
to the avatar creator when it became ready. Google
did not like this and this is unnecessarily complicated
anyway.

Now we just create a disk directly from the guest image.
This leaves less room for google's shenanigans and
is better in every way. I have yet to see a 404 or
any resources mysteriously disappear in this new world.
So I think this change will fix the 404 problem

Testing: encrypted a lot of images