Handle import errors inside subcommand module
When importing a subcommand module, check ImportError.message and raise
the exception if the failed import is not the one we're expecting.

History: a subcommand import can fail for two reasons:

1) The subcommand module is missing
2) The subcommand module imports another module that's missing

We were not differentiating between these two cases, which was causing
subcommands to be ignored when a dependency was not installed.