Handle edge case errors when calling the AWS API
Handle two obscure errors that occasionally pop up in the field.  I
believe that the AWS service returns 503 when the request rate limit has
been exceeded for a given account.  I'm not sure what causes
ssl.SSLError, but it has been reported a couple of times.

To handle these errors, we:

* Check for these cases in BotoRetryExceptionChecker.
* Wrap every AWS call with retry_boto()

Also, make BotoRetryExceptionChecker.error_code_regexp an optional
argument.  If the regexp isn't specified, we check for the two cases
above.

Change "raise e" to "raise" in util.retry(), so that the original
traceback is preserved.