Removes username and password authentication.
Rather than use username and password to
authenticate with yeti a token created with
brkt make-token is now used. This token is
signed by a users private key whose corresponding
public key has been registered with yeti. This
allows us to verify the authenticity of the token
and its claims without a uname and password.

The --token arg is now used to pass a
token through userdata to the encryptor for
both gce and aws (though it is optional in aws).

Testing: encrypted an image in gce