Adds brkt share-logs subcommand
This subcommand allows users to share a snapshot of the Bracket
system log volume from an instance in the customer AWS account
with the Bracket AWS account.

This is intended to be used by customers to aid Bracket debugging
in cases where logs are not reaching the Bracket servers (e.g.
cases with network connecitivity issues from an instance).

Sample output:
./brkt share-logs --region us-west-2 i-c8b93215
19:22:56 Creating snapshot snap-d9a3ba81 of log volume for instance
i-c8b93215
19:23:58 snap-d9a3ba81: 90%
19:24:44 Sharing snapshot snap-d9a3ba81 with AWS account 164337164081
snap-d9a3ba81

Tested by sharing the log volume of a running instance.