Handle eventual consistency errors with temporary security group
Handle eventual consistency issues with the temporary security group
that allows brkt-cli to poll metavisor for encryption and update
progress.

We create the security group and then immediately attempt to launch the
encryptor instance with it, which occasionally causes AWS to return an
InvalidGroup.NotFound error code.  Detect this case and wrap the
run_instace() call in a retry.

Add unit tests that verify that we handle the error.  Implement
DummyAWSService.retry(), so that unit tests can test retry behavior.

Call util.sleep() instead of time.sleep() in util.retry(), so that unit
tests don't sleep.