Snapshots log volume for failed encryption
Adds a flag to specify whether to save logs from encryptor
instance in a snapshot. This is by default True.

Allows `brkt share-logs` to accept a --snapshot-id.

Use case for failed encryption is customer runs an encryption
and if it fails, the snapshot of the log volume is taken and
the brkt share-logs command is logged for users if they then
choose to share with Bracket support.