Add support for specifying options in a config file
This commit adds infrastructure to allow subcommands to read default
values from a config file stored in ~/.brkt/config. In order to use
the new infrastructure, subcommands must implement the setup_config()
method. In it, they should register any options they want to pull
from the config.

A new sub-command, "config", has been added to the cli. This lets
users list, set, get, and unset options in the config. It is
modeled after "kubectl config".

Finally, the first consumer of the config infrastructure is the
"encrypt-gce-image" sub-command. The "project", "network", and
"zone" options may be specified via the config.