Support specifying the Yeti environment by its domain
Add a new hidden --service-domain option that specifies the Yeti
environment that the metavisor (and soon, the CLI) will talk to.
When a domain is specified (e.g. stage.mgmt.brkt.com), we infer the
following hosts and ports:

* internal API (RPC): yetiapi.<domain>:443
* HSM proxy: hsmproxy.<domain>:443
* public API: api.<domain>:80

Always send the Yeti environment to metavisor, so that behavior is
clearly defined and does not differ between CSPs.