Validate JWT before attempting encryption
When the user specifies --token, check the JWT against the public JWK
endpoint before attempting encryption.  This causes the encryption
process to fail earlier and more intuitively when the public key is not
associated with a Yeti customer.

Fix a bug in BracketEnvironment initialization.  The default https port
is 443, not 80.