Don't require brkt-env when launching a GCE instance
Split BracketEnvironment parsing out from the rest of the values passed
when constructing InstanceConfig.  This allows callsites to determine
whether to specify BracketEnvironment or not.

Background: in commit bfe70c372f95730964a0f9c3366aa94a8806cf64 we
started always sending the Yeti environment to Metavisor.  This created
problems when launching a GCE image.  If the user didn't specify the
environment on the command line, we'd send the prod endpoints, which
would inadvertently override the Metavisor defaults.