Add support for user-specified environments stored in config
This commit makes environment a first class concept in the CLI. Environments are
defined as the set of Yeti backends that the CLI/MV talk to. Users declare new
environments via the `set-env` config subcommand. For example, to define the
`foo.com` environment, one does:

    > brkt config set-env foo --service-domain foo.com

Use the `use-env` config subcommand to switch between environments:

    > brkt config use-env stage

Use the `list-envs` config subcommand to list all the declared environments. A
leading `*` indicates the current environment.

    > ./brkt config list-envs
      brkt-hosted
    * foo
    >

The `brkt-hosted` environment corresponds to our hosted production environment.
It is the default environment and may not be changed by users.

Use the `get-env` config subcommand to get the details of an environment:

    > ./brkt config get-env foo
    api-server=yetiapi.foo.com:443
    key-server=hsmproxy.foo.com:443
    network-server=network.foo.com:443
    public-api-server=api.foo.com:443
    >

Finally, use the `unset-env` config subcommand to delete an environment:

    > ./brkt config unset-env foo