Print only fatal messages from boto
Boto tries to fetch the AWS credentials from metadata service before
doing a S3 download. If it does not get those credentials (which are
not present in ESX), it prints a confusing info message which looks
like an error. Hence, making the change to print only fatal messages
from boto.