Update the AWS command line spec
Update the AWS command line spec to follow our new convention.  Move the
following commands:

* encrypt-ami -> aws encrypt
* update-encrypted-ami -> aws update
* diag -> aws diag
* share-logs -> aws share-logs

Hide the old commands and log a deprecation warning when they are used.
Refactor the AWS module, so that the run() functions in both the old
and new commands call the same code.

The GCE and VMware commands will be updated in separate commits.