Update GCE command line spec
Update the GCE command line spec to follow our new convention. Move the following commands:

 - encrypt-gce-image -> gce encrypt-image
 - update-gce-image -> gce update-image
 - launch-gce-image -> gce launch-image

Hide the old commands and log a deprecation warning when they are used.
Refactor the GCE module so that the run() functions in both the old
and new commands call the same code.
Migrate the GCE config options based on the new command convention.

The AWS and ESX commands will be udpated in separate commits.