Parse GCE config options correctly
The GCE config options were being references using the old names
because of which they were not being parsed (and utilized) correctly
with the GCE specific brkt-cli commands. With this change, even after
the config options are migrated to the new format, the config
options would be utilized both with the new command syntax (for
e.g. *brkt gce encrypt*) as well as the old command syntax (for
e.g. *brkt encrypt-gce-image*)